Fixed issue with certain OpenGL blend functions in HDR
From what I could find, this behavior is poorly documented. glBlendFunc() states that "blending operates with imprecise integer color values" and glBlendEquation() states that "The results of these equations are clamped to the range [0,1]."

Since alpha values are still between 0 and 1 in HDR mode, any blend using ONLY alpha values functions normally. However, blends using any R, G, or B values > 1 seem to overflow to infinity (and are NOT clamped to between 0 and 1). This overflow creates a number of visual artifacts that trash any blend operations occurring afterwards.

https://stackoverflow.com/questions/31960193/why-does-a-black-source-color-with-glblendfuncgl-src-alpha-gl-one-still-caus

This SO user seems to have encountered the same issue. Despite a comment stating that "OpenGL 3.0 and later should support all blending operations on float buffers" this does not seem to be the case (we're using OpenGL 3.3 Core).